backend/requirements.txt
psycopg2==2.9.5
Django==3.2.5
pandas==1.2.2
djangorestframework==3.14.0
python-decouple==3.6
